--- dictonaries.py
@@ -1,116 +1,0 @@
-countries = {
-    'EU27_2020': "European Union",
-    'BE': 'Belgium',
-    'BG': 'Bulgaria',
-    'CZ': 'Czechia',
-    'DK': 'Denmark',
-    'DE': 'Germany', 
-    'EE': 'Estonia',
-    'IE': 'Ireland',
-    'EL': 'Greece',
-    'ES': 'Spain',
-    'FR': 'France',
-    'HR': 'Croatia',
-    'IT': 'Italy',
-    'CY': 'Cyprus',
-    'LV': 'Latvia',
-    'LT': 'Lithuania',
-    'LU': 'Luxembourg',
-    'HU': 'Hungary',
-    'MT': 'Malta',
-    'NL': 'Netherlands',
-    'AT': 'Austria',
-    'PL': 'Poland',
-    'PT': 'Portugal',
-    'RO': 'Romania',
-    'SI': 'Slovenia',
-    'SK': 'Slovakia',
-    'FI': 'Finland',
-    'SE': 'Sweden',
-    'IS': 'Iceland',
-    'NO': 'Norway',
-    'CH': 'Switzerland' 
-}
-
-categories ={
-'TOTAL':'Total',
-'GF01':'General public services',
-'GF0101':'Executive and legislative organs, financial and fiscal affairs, external affairs' ,
-'GF0102':'Foreign economic aid',
-'GF0103':'General services',
-'GF0104':'Basic research',
-'GF0105':'R&D General public services',
-'GF0106':'General public services n.e.c.',
-'GF0107':'Public debt transactions',
-'GF0108':'Transfers of a general character between different levels of government',
-'GF02':'Defence',
-'GF0201':'Military defence',
-'GF0202':'Civil defence',
-'GF0203':'Foreign military aid',
-'GF0204':'R&D Defence',
-'GF0205':'Defence n.e.c.',
-'GF03':'Public order and safety',
-'GF0301':'Police services',
-'GF0302':'Fire-protection services',
-'GF0303':'Law courts ',
-'GF0304':'Prisons',
-'GF0305':'R&D Public order and safety',
-'GF0306':'Public order and safety n.e.c.',
-'GF04':'Economic affairs',
-'GF0401':'General economic, commercial and labour affairs',
-'GF0402':'Agriculture, forestry, fishing and hunting',
-'GF0403':'Fuel and energy',
-'GF0404':'Mining, manufacturing and construction',
-'GF0405':'Transport',
-'GF0406':'Communication',
-'GF0407':'Other industries',
-'GF0408':'R&D Economic affairs',
-'GF0409':'Economic affairs n.e.c.',
-'GF05' :'Environmental protection',
-'GF0501': 'Waste management',
-'GF0502':'Waste water management',
-'GF0503': 'Pollution abatement',
-'GF0504':'Protection of biodiversity and landscape ',
-'GF0505':'R&D Environmental protection',
-'GF0506':'Environmental protection n.e.c.',
-'GF06':'Housing and community amenities',
-'GF0601':'Housing development',
-'GF0602':'Community development',
-'GF0603':'Water supply',
-'GF0604':'Street lighting',
-'GF0605':'R&D Housing and community amenities',
-'GF0606':'Housing and community amenities n.e.c.',
-'GF07' :'Health',
-'GF0701':'Medical products, appliances and equipment',
-'GF0702':'Outpatient services',
-'GF0703':'Hospital services',
-'GF0704':'Public health services',
-'GF0705':'R&D Health',
-'GF0706':'Health n.e.c.',
-'GF08':'Recreation, culture and religion' ,
-'GF0801':'Recreational and sporting services',
-'GF0802':'Cultural services',
-'GF0803':'Broadcasting and publishing services',
-'GF0804':'Religious and other community services',
-'GF0805':'R&D Recreation, culture and religion',
-'GF0806':'Recreation, culture and religion n.e.c.',
-'GF09':'Education',
-'GF0901':'Pre-primary and primary education',
-'GF0902':'Secondary education',
-'GF0903':'Post-secondary non-tertiary education ',
-'GF0904':'Tertiary education',
-'GF0905':'Education not definable by level',
-'GF0906':'Subsidiary services to education',
-'GF0907':'R&D Education',
-'GF0908':'Education n.e.c.',
-'GF10':'Social protection',
-'GF1001':'Sickness and disability',
-'GF1002':'Old age',
-'GF1003':'Survivors',
-'GF1004':'Family and children',
-'GF1005':'Unemployment',
-'GF1006':'Housing',
-'GF1007':'Social exclusion n.e.c.',
-'GF1008':'R&D Social protection',
-'GF1009':'Social protection n.e.c.' 
-}
