+++ pages/3_📈_Find Max.py
@@ -1,11 +1,13 @@
 import streamlit as st
 import pandas as pd
-from dictonaries import categories
-from clean import cleaning
-from functions import find_max
-millions = pd.read_csv('MIL_coun31_12-20_cat80.csv')
-percents = pd.read_csv('GDP_coun31_12-20_cat80.csv')
-TotalGDP = pd.read_csv('mil_GDP_TOTAL_31_12-20.csv')
+from tools.dictonaries import categories
+from tools.clean import cleaning
+from tools.functions import find_max
+
+percents = pd.read_csv('datasets\GDP_coun31_12-20_cat80.csv')
+millions = pd.read_csv('datasets\MIL_coun31_12-20_cat80.csv')
+TotalGDP = pd.read_csv('datasets\mil_GDP_TOTAL_31_12-20.csv')
+
 df_percs = cleaning(percents)
 df_money = cleaning(millions)
 e = NameError('Please upload a file from sidebar')
